This is modification of Nelson Rouque's php implementation of the automated operation span task (aospan)
This version of the task is shortened from about 20 minutes to 8
minutes. It follows guidelines from the psychological
literature that show that truncated WMC tasks retain reliability and
validity while allowing working memory to be
assessed quickly and allowing researchers to measure other constructs
within the natural limits of testing time and the
natural limits on test-taker’s patience.

Design guideliness were obtained from:
https://link.springer.com/article/10.3758/s13428-014-0543-2

Compared to the full aospan, this task presents 2 iterations of set
sizes 4,5, and 6 for a total of 30 trials (compared to 75 in the
original task).

In addition, a few bug fixes were implemented, described in README.